Add offset alignment to photots
