+++ src/components/App/App.css
@@ -6,40 +6,13 @@
 }
 
 .main {
-  /* background: #131313; */
-}
-
-.container {
-  max-width: 1320px;
-  margin: 0 auto;
+  width: 100%;
+  overflow-x: hidden;
 }
 
 .img-container {
   margin: 10% 0;
   height: 70%;
-  position: relative;
-}
-
-.img-container_:nth-child(even) {
-  margin-right: 0;
-  margin-left: auto;
-}
-
-img {
-  height: 60%;
-  width: 60%;
-}
-
-/* . mask {
-
-} */
-
-.mask {
-  position: absolute;
-  /* border: 5px red solid; */
   width: 100%;
-  height: 100%;
-  background: #131313;
-  overflow: hidden;
-  /* visibility: hidden; */
+  position: relative;
 }
